+++ packages/client/views/TagVideos/action/make-query-tag-action.ts
@@ -1,0 +1,8 @@
+import { Action } from 'react-fetching-library';
+
+import { endpoint } from '../../../constants';
+
+export const makeQueryTagAction: Action = (id: number) => ({
+  method: 'GET',
+  endpoint: `${process.env.API_URL_HOST}${endpoint.tags}/${id}`,
+});
