+++ packages/client/views/TagVideos/action/make-query-tag-videos-action.ts
@@ -2,12 +2,7 @@ import { Action } from 'react-fetching-library';
 
 import { endpoint } from '../../../constants';
 
-export const getTagAction: Action = (id: number) => ({
-  method: 'GET',
-  endpoint: `${process.env.API_URL_HOST}${endpoint.tags}/${id}`,
-});
-
-export const getTagVideosAction: Action = (
+export const makeQueryTagVideosAction: Action = (
   id: number,
   page: number,
   sort: string,
